+++ lm_eval/tasks/sympy_math.py
@@ -20,8 +20,7 @@ import inspect
 import lm_eval.datasets.hendrycks_math.hendrycks_math
 from lm_eval.metrics import mean
 from lm_eval.base import Task, rf
-from lm_eval.utils import MajorityVotingMixin, SymbolicMathMixin
-from lm_eval.tasks.math_tasks import SymbolicMathTask
+from lm_eval.mixins import MajorityVotingMixin, SymbolicMathMixin
 
 
 CODE_PROMPT=r"""Problem:
